requirements.txt
python-telegram-bot==21.6
tzdata==2024.1
